refactor: reward module to common.
