+++ src/common/reward/dto/reward.dto.ts
