+++ src/common/reward/models/reward.entity.ts
