+++ src/substrate/substrate.controller.ts
@@ -2,8 +2,8 @@ import { Body, Controller, Post, Res, Headers, UseInterceptors } from '@nestjs/c
 import { SubstrateService } from './substrate.service';
 import { Response } from 'express';
 import { ApiProperty } from '@nestjs/swagger';
-import { RewardService } from '../reward/reward.service';
-import { RewardDto } from '../reward/dto/reward.dto';
+import { RewardService } from '../common/reward/reward.service';
+import { RewardDto } from '../common/reward/dto/reward.dto';
 import { SentryInterceptor } from '../common';
 
 export type RegistrationRole = 'lab' | 'doctor' | 'hospital';
