+++ src/substrate/substrate.module.ts
@@ -3,7 +3,7 @@ import { SubstrateService } from './substrate.service';
 import { SubstrateController } from './substrate.controller';
 import { EscrowModule } from '../endpoints/escrow/escrow.module';
 import { TransactionLoggingModule } from '../transaction-logging/transaction-logging.module';
-import { RewardModule } from '../reward/reward.module';
+import { RewardModule } from '../common/reward/reward.module';
 import { MailModule } from '../common/mailer/mailer.module';
 import { LocationModule } from '../endpoints/location/location.module';
 import { DebioConversionModule } from '../common/debio-conversion/debio-conversion.module';
