+++ test/unit/endpoints/verification/verification.service.spec.ts
@@ -3,9 +3,9 @@ import { dateTimeProxyMockFactory, MockType } from '../../mock';
 import { when } from 'jest-when';
 import { DateTimeProxy } from '../../../../src/common/date-time';
 import { SubstrateService } from '../../../../src/substrate/substrate.service';
-import { RewardService } from '../../../../src/reward/reward.service';
+import { RewardService } from '../../../../src/common/reward/reward.service';
 import { VerificationService } from '../../../../src/endpoints/verification/verification.service';
-import { RewardDto } from '../../../../src/reward/dto/reward.dto';
+import { RewardDto } from '../../../../src/common/reward/dto/reward.dto';
 
 describe('Verification Service Unit Tests', () => {
   let verificationService: VerificationService;
